nvperf: new program.
Allow changing NVidia powermizer performance level: adaptive, or
maximum. The program is based on a state file, containing a counter that
is atomically incremented when asking for maximum performance, and
decremented when asking for adaptive (>= 0). This allow multiple
concurrent programs (game, media player, ...) to ask for maximum
performance at the beginning of their operation, and adaptive
performance when their done, and the level will automatically switch
back to adaptive when the last program is done (counter reach zero).